Use CMAKE_Fortran_COMPILER_ID instead of CMAKE_Fortran_COMPILER
For checking whether the Intel compiler is being used, employ the
CMake variable CMAKE_Fortran_COMPILER_ID rather than the original
CMAKE_Fortran_COMPILER.

The original code did not support using the "mpifort" compiler
wrapper with (for instance) gfortran, because "mpifort" matched
the pattern "ifort", which resulted in addition of compiler flags
not supported by gfortran.

On the contrary, for Intel compiler, CMAKE_Fortran_COMPILER_ID is
set to "Intel", regardless whether it is used directly or inside
a MPI wrapper. Similarly, it returns "GNU" for gfortran, whether
used directly or via "mpifort".